Added blocks to libary. Not yet implemented successfully in Main.slx simulation. 1. Sequential Motor Controller This block is currently in a development state, though it can be quickly and easily converted to flight state. It takes as inputs the state and spin rate vector. It outputs an impulse signal to trigger an unreeling event. The magnitude of the impulse is the length by which the tether is unreeled. Therefore, the output isn't physical reeling rates, but rather signals.
2. Update Tether properties
This block will be part of the dynamics model. It accepts the unreeling signals from the Sequential Motor Controller block and outputs the current tether length as a signal. This output is used by the tether models to update the physical tether length. It also calls a MATLAB function to update the mechanical properties of the tether links using set_param() at each unreeling event.